src/main/java/com/example/garage/UserService/EncryptionService.java
package com.example.garage.UserService;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.security.crypto.bcrypt.BCrypt;
import org.springframework.stereotype.Service;

@Service
public class EncryptionService {

    @Value("${encryption.salt.rounds}")
    private int saltRounds;

    public String encryptPassword(String password) {
        return BCrypt.hashpw(password, BCrypt.gensalt(saltRounds));
    }

    public boolean verifyPassword(String password, String hash) {
        return BCrypt.checkpw(password, hash);
    }
}
